Update README.md: add JupyterLab client
Thank you for maintaining this language server! Just wanted to let you know that we support JavaScript and TypeScript using your package, and add a shameless promotion to help users who may look for a client with JS/TS support.

While we may be late to the party, our repo has a link to binder which allows running the JupyterLab with LSP integration in the browser without installing anything - might be of use to some "searching" users.